requirements.txt
Flask==3.1.0
Flask-WTF==1.2.2
SQLAlchemy==2.0.16
Flask-SQLAlchemy==3.1.0
Werkzeug==3.1.3
python-dotenv==1.1.0
psycopg2-binary==2.9.6